Multiple endpoints in config. Some refactor
