+++ ton_client/src/net/server_link.rs
@@ -17,6 +17,7 @@ use crate::net::gql::{GraphQLOperation, GraphQLOperationEvent, OrderBy, PostRequ
 use crate::net::server_info::ServerInfo;
 use crate::net::websocket_link::WebsocketLink;
 use crate::net::{Error, NetworkConfig};
+use super::websocket_link::WsConfig;
 use futures::{Future, Stream, StreamExt};
 use serde_json::Value;
 use std::collections::HashMap;
@@ -33,6 +34,7 @@ pub(crate) struct Subscription {
 
 pub(crate) struct ServerLink {
     config: NetworkConfig,
+    endpoints: Vec<String>,
     client_env: Arc<ClientEnv>,
     suspended: AtomicBool,
     server_info: tokio::sync::RwLock<Option<ServerInfo>>,
@@ -42,15 +44,23 @@ pub(crate) struct ServerLink {
 }
 
 impl ServerLink {
-    pub fn new(config: NetworkConfig, client_env: Arc<ClientEnv>) -> Self {
-        ServerLink {
+    pub fn new(config: NetworkConfig, client_env: Arc<ClientEnv>) -> ClientResult<Self> {
+        let endpoints = config.endpoints.clone()
+            .or(config.server_address.clone().map(|address| vec![address]))
+            .ok_or(crate::client::Error::net_module_not_init())?;
+        if endpoints.len() == 0 {
+            return Err(crate::client::Error::net_module_not_init());
+        }
+
+        Ok(ServerLink {
             config: config.clone(),
+            endpoints,
             client_env: client_env.clone(),
             suspended: AtomicBool::new(false),
             query_url: std::sync::RwLock::new(None),
             server_info: tokio::sync::RwLock::new(None),
-            websocket_link: WebsocketLink::new(config, client_env.clone()),
-        }
+            websocket_link: WebsocketLink::new(client_env.clone()),
+        })
     }
 
     async fn query_by_url(&self, address: &str, query: &str) -> ClientResult<Value> {
@@ -96,11 +106,27 @@ impl ServerLink {
     }
 
     async fn init(&self, config: &NetworkConfig) -> ClientResult<ServerInfo> {
-        let queries_server = ServerInfo::expand_address(&config.server_address);
-        let server_info = ServerInfo::fetch(self.client_env.clone(), &queries_server).await?;
+        let mut futures = vec![];
+        for address in &self.endpoints {
+            let queries_server = ServerInfo::expand_address(&address);
+            futures.push(Box::pin(async move { 
+                ServerInfo::fetch(self.client_env.clone(), &queries_server).await
+            }));
+        }
+        
+        let mut server_info = Err(crate::client::Error::net_module_not_init());
+        while futures.len() != 0 {
+            let (result, _, remain_futures) = futures::future::select_all(futures).await;
+            futures = remain_futures;
+            server_info = result;
+            if server_info.is_ok() {
+                break;
+            }
+        }
+        let server_info = server_info?;
 
         if server_info.server_version.supports_time {
-            self.check_time_delta(&queries_server, config).await?;
+            self.check_time_delta(&server_info.query_url, config).await?;
         }
 
         Ok(server_info)
@@ -121,6 +147,12 @@ impl ServerLink {
         }
 
         let inited_data = self.init(&self.config).await?;
+
+        self.websocket_link.set_config(WsConfig {
+            url: inited_data.subscription_url.clone(),
+            access_key: self.config.access_key.clone()
+        }).await;
+
         *self.query_url.write().unwrap() = Some(inited_data.query_url.clone());
         *data = Some(inited_data);
 
@@ -131,8 +163,8 @@ impl ServerLink {
         &self.config
     }
 
-    pub fn config_server(&self) -> &str {
-        &self.config.server_address
+    pub fn config_servers(&self) -> &[String] {
+        &self.endpoints
     }
 
     pub fn query_url(&self) -> Option<String> {
@@ -146,6 +178,8 @@ impl ServerLink {
         filter: &Value,
         fields: &str,
     ) -> ClientResult<Subscription> {
+        self.ensure_info().await?;
+        
         let event_receiver = self
             .websocket_link
             .start_operation(GraphQLOperation::subscription(table, filter, fields))
